+++ fastlib/trunk/contrib/dongryel/thesis_research/mlpack/two_point/two_point_result.h
@@ -111,7 +111,7 @@ namespace mlpack {
       } // ApplyProbabalisticDelta
       
       
-      void Seed(int qpoint_index, double initial_pruned_in) {
+      void Seed(int qpoint_index, unsigned long initial_pruned_in) {
         
       }
